# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PROSPECTIVE Web API",
    "description": "REST + WebSocket API for cerebral aneurysm segmentation, detection, morphometry, perforator risk assessment and stent planning.\n\nAll linear dimensions are in **millimetres (mm)**. All HU values follow the standard Hounsfield scale.",
    "contact": {
      "name": "SkullApp — Laboratorio de Imagen Médica",
      "email": "ingprospective@skullapp.tech"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/clips": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Get clip device library",
        "description": "Returns the full surgical clip library (40+ models across Yasargil, Sugita, Aesculap and Codman systems). Each entry includes blade length, shape, closing force, and the required applier.",
        "operationId": "get_clip_library_api_clips_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClipLibraryItem"
                  },
                  "type": "array",
                  "title": "Response Get Clip Library Api Clips Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/clips/recommendations/{session_id}": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Get clip recommendations from the Clip Recommender assistant",
        "description": "Scores all clips in the catalogue against the patient's neck diameter and aspect ratio (loaded from session morphometry). Returns up to 8 ranked recommendations.\n\nScoring weights: coverage ratio 45%, shape fit 40%, closing force 15%.",
        "operationId": "get_clip_recommendations_api_clips_recommendations__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClipRecommendation"
                  },
                  "title": "Response Get Clip Recommendations Api Clips Recommendations  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/clips/custom/{session_id}": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Upload a custom clip mesh (STL/OBJ)",
        "description": "Imports a user-supplied clip geometry (STL or OBJ), centres it at the origin and stores it for placement. Returns a `custom:*` clip_id usable in the clip plan alongside catalogue clips.",
        "operationId": "upload_custom_clip_api_clips_custom__session_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_custom_clip_api_clips_custom__session_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomClipInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Custom clips imported in this session",
        "description": "The imported clips survive in the session directory, but the browser forgets them on resume, so the dropdown lost geometry that was still on disk. This restores the list.",
        "operationId": "list_custom_clips_api_clips_custom__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomClipInfo"
                  },
                  "title": "Response List Custom Clips Api Clips Custom  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/clips/plan": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Compute clip placement plan",
        "description": "Builds a real 3D mesh of each placed clip, checks collision against the segmented vessel with vtkCollisionDetectionFilter, and estimates neck coverage from the intersection of the clips with the neck plane. Returns the combined clip mesh URL (.vtp) for the 3D viewer.",
        "operationId": "plan_clips_api_clips_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipPlanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/clips/custom/{session_id}/{clip_index}": {
      "delete": {
        "tags": [
          "clips"
        ],
        "summary": "Remove an imported custom clip",
        "description": "Deletes the imported geometry and its catalogue entry. Without it a mis-imported file stayed in the dropdown for the rest of the session. Returns the remaining custom clips.\n\nClips already placed in the plan are untouched — clear those from the devices step.",
        "operationId": "delete_custom_clip_api_clips_custom__session_id___clip_index__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "clip_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clip Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomClipInfo"
                  },
                  "title": "Response Delete Custom Clip Api Clips Custom  Session Id   Clip Index  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_custom_clip_api_clips_custom__session_id__post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_custom_clip_api_clips_custom__session_id__post"
      },
...
      "ClipLibraryItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Commercial name (e.g. 'Yasargil FT740T')"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "length_mm": {
            "type": "number",
            "title": "Length Mm",
            "description": "Blade length (mm). Typical: 5–20 mm"
          },
          "angle_deg": {
            "type": "number",
            "title": "Angle Deg",
            "description": "Clip angle (0 = straight, 90 = right-angle)",
            "default": 0.0
          },
          "is_fenestrated": {
            "type": "boolean",
            "title": "Is Fenestrated",
            "description": "True for fenestrated clips used on vessel bifurcations",
            "default": false
          },
          "closing_force_g": {
            "type": "number",
            "title": "Closing Force G",
            "description": "Closing force in grams (typical: 80–200 g)"
          },
          "compatible_applier": {
            "type": "string",
            "title": "Compatible Applier",
            "description": "Required applier instrument model"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "manufacturer",
          "length_mm",
          "closing_force_g",
          "compatible_applier"
        ],
        "title": "ClipLibraryItem",
        "description": "A surgical clip model from the device library."
      },
      "ClipPlacement": {
        "properties": {
          "clip_id": {
            "type": "string",
            "title": "Clip Id",
            "description": "ID of the clip from the library"
          },
          "position": {
            "$ref": "#/components/schemas/Position3D",
            "description": "Clip jaw centre in patient space (mm)"
          },
          "normal": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Normal",
            "description": "Clip blade normal vector [nx, ny, nz] — perpendicular to blade"
          },
          "rotation_deg": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Rotation Deg",
            "description": "Rotation around the normal axis (degrees)",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "clip_id",
          "position",
          "normal"
        ],
        "title": "ClipPlacement",
        "description": "One clip placed on the aneurysm neck."
      },
      "ClipPlanRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "placements": {
            "items": {
              "$ref": "#/components/schemas/ClipPlacement"
            },
            "type": "array",
            "title": "Placements",
            "description": "All clips to place (replaces the current plan)"
          },
          "trajectory_entry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position3D"
              },
              {
                "type": "null"
              }
            ],
            "description": "Surgical approach entry point in patient space (mm)"
          },
          "trajectory_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position3D"
              },
              {
                "type": "null"
              }
            ],
            "description": "Surgical approach target point in patient space (mm)"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "placements"
        ],
        "title": "ClipPlanRequest",
        "description": "Request to add or update a clip in the planning."
      },
      "ClipPlanResult": {
        "properties": {
          "clips_mesh_url": {
            "type": "string",
            "title": "Clips Mesh Url",
            "description": "URL of the combined clips mesh (.vtp) for 3D display"
          },
          "trajectory_mesh_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trajectory Mesh Url",
            "description": "URL of the trajectory cylinder mesh (.vtp)"
          },
          "neck_coverage_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Neck Coverage Pct",
            "description": "Percentage of the neck cross-section occluded by the clips"
          },
          "collision_detected": {
            "type": "boolean",
            "title": "Collision Detected",
            "description": "True if any clip intersects a vessel wall or another clip",
            "default": false
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning",
            "description": "Warning when collision or poor coverage is detected"
          }
        },
        "type": "object",
        "required": [
          "clips_mesh_url",
          "neck_coverage_pct"
        ],
        "title": "ClipPlanResult",
        "description": "Result of a clip placement plan."
      },
      "ClipRecommendation": {
        "properties": {
          "clip_id": {
            "type": "string",
            "title": "Clip Id"
          },
          "clip_name": {
            "type": "string",
            "title": "Clip Name"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Recommendation score (higher = better fit)"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "One-line clinical rationale for this clip"
          },
          "suggested_placement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClipPlacement"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pre-computed suggested placement, if available"
          }
        },
        "type": "object",
        "required": [
          "clip_id",
          "clip_name",
          "score",
          "reason"
        ],
        "title": "ClipRecommendation",
        "description": "Clip recommendation from the Clip Recommender assistant."
      },
...
      "CustomClipInfo": {
        "properties": {
          "clip_id": {
            "type": "string",
            "title": "Clip Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "clip_id",
          "name"
        ],
        "title": "CustomClipInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Position3D": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X",
            "description": "X coordinate in mm (patient space)"
          },
          "y": {
            "type": "number",
            "title": "Y",
            "description": "Y coordinate in mm (patient space)"
          },
          "z": {
            "type": "number",
            "title": "Z",
            "description": "Z coordinate in mm (patient space)"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "z"
        ],
        "title": "Position3D"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}